Update index.py - High score saving
Solved the issue where the file didn't previously exist, the program now checks to see if the text file exists and creates a new text file if it does not. If the file does exist, the text file will not be overwritten by this code. The only time it will be overwritten is when the user gets a new high score.